[Feature]Implement dump results (#8814)
* Implement dump results

* support xxxevaluator

* update

* reuse func
