+++ mmdet/engine/hooks/__init__.py
@@ -5,11 +5,12 @@ from .memory_profiler_hook import MemoryProfilerHook
 from .num_class_check_hook import NumClassCheckHook
 from .set_epoch_info_hook import SetEpochInfoHook
 from .sync_norm_hook import SyncNormHook
+from .utils import trigger_visualization_hook
 from .visualization_hook import DetVisualizationHook
 from .yolox_mode_switch_hook import YOLOXModeSwitchHook
 
 __all__ = [
     'YOLOXModeSwitchHook', 'SyncNormHook', 'CheckInvalidLossHook',
     'SetEpochInfoHook', 'MemoryProfilerHook', 'DetVisualizationHook',
-    'NumClassCheckHook', 'MeanTeacherHook'
+    'NumClassCheckHook', 'MeanTeacherHook', 'trigger_visualization_hook'
 ]
